+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Added `plan` function which takes a schema object of functions to apply to a matching object of data
 - Added `gt` function which compares the passed in values to determine if one is greater than the other
 - Added `gte` function which compares the passed in values to determine if one is greater than or equal to the other
+- Added `lt` function which compares the passed in values to determine if one is less than the other
 
 ### Improved